requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
flask==3.0.0
flask-cors==4.0.0
schedule==1.2.0
pandas==2.1.3
numpy==1.26.2
plotly==5.18.0
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3